requirements.txt
openslide-python
numpy
pillow
matplotlib
pandas
opencv-python
pyyaml
tqdm
torch
torchvision
scikit-image
